Resolve "audit public v2 docs ready for go-live"

See merge request ukfast/api.ukfast/ecloud!703

+++ app/Models/V2/BillingMetric.php
@@ -46,7 +46,7 @@ class BillingMetric extends Model implements Filterable, Sortable
 
         $this->casts = [
             'price' => 'float',
-            'value' => 'integer'
+            'value' => 'float'
         ];
 
         parent::__construct($attributes);
